requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LlamaIndex
llama-index==0.9.14
llama-index-embeddings-openai==0.1.5
llama-index-llms-openai==0.1.5
llama-index-vector-stores-postgres==0.1.3

# Database
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
testcontainers==3.7.1
